Refactored menu to separate component
